requirements.txt
fastapi
uvicorn
python-dotenv
pandas
chromadb
sentence-transformers
langchain
langchain-google-genai